pyproject.toml
[project]
name = "tear-cls"
version = "0.1.0"
description = "Fine-tune a vision LLM on AFM tear-film images for diagnosis classification"
requires-python = ">=3.12,<3.13"
dependencies = [
    "torch==2.5.1",
    "torchvision",
    "transformers<4.50",
    "trl>=0.12.0",
    "peft>=0.14.0",
    "accelerate>=1.0.0",
    "bitsandbytes>=0.44.0",
    "datasets>=3.0.0",
    "pillow>=10.0",
    "scikit-learn>=1.5",
    "open-clip-torch>=2.30",
    "tqdm",
    "torchao<0.12",
    "wandb>=0.26.0",
    "albumentations>=1.4",
    "opencv-python-headless>=4.9",
]

[tool.uv]
# Pin torch to cu121 wheels — works with the CUDA 13 driver via backward compat
[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
